new completely refactored version of msm_basic subpackage
* improved performance and memory consumption due to dataset segment
being processed like a stream of ion images
* both dataset and centroids split into segments
* works only locally (no S3 support yet)
